+++ src/main/java/com/github/pagehelper/dialect/AbstractHelperDialect.java
@@ -188,7 +188,7 @@ public abstract class AbstractHelperDialect extends AbstractDialect implements C
         String orderBy = page.getOrderBy();
         if (StringUtil.isNotEmpty(orderBy)) {
             pageKey.update(orderBy);
-            sql = OrderByParser.converToOrderBySql(sql, orderBy, jSqlParser);
+            sql = OrderByParser.converToOrderBySql(sql, orderBy);
         }
         if (page.isOrderByOnly()) {
             return sql;